feat: Add SVG assets, update dependencies, and configure UI
This commit adds two new SVG assets, `bg2.svg` and `wavebg.svg`, to the `smartcanteen/assets` directory. It also updates the `react-native-paper` dependency to version `5.12.3` and adds the `react-native-svg` dependency with version `15.2.0` to the `package.json` file. These changes are necessary for improving the visual appearance of the app and enabling SVG support.

Additionally, the UI is configured and a bottom app bar is built, enhancing the user experience and providing better navigation within the app.